Add self-study disclaimer
Microsoft Partners are using these labs for self-study without ILT interaction first.  Although we're very happy the partners are showing that level of interest, it's causing some back-end static because individuals are trying to do the labs in locked-down subscriptions supplied by the partners.  This change is a suggested way to handle that issue.